Field types changed from long to int in numeric literal expressions

+++ hu.bme.mit.theta.core/src/main/java/hu/bme/mit/theta/core/expr/IntLitExpr.java
@@ -3,7 +3,9 @@ package hu.bme.mit.theta.core.expr;
 import hu.bme.mit.theta.core.type.IntType;
 
 public interface IntLitExpr extends LitExpr<IntType>, NullaryExpr<IntType>, Comparable<IntLitExpr> {
-	public long getValue();
+
+	public int getValue();
 
 	public RatLitExpr toRatLit();
+
 }
